Changed every point to float

+++ Graphics_Lib/Shapes/circle.hpp
@@ -7,12 +7,12 @@
 
 class Circle : public Shape {
 public:
-	Circle(const V2d<int> & p, int r) 
+	Circle(const V2d<float> & p, int r) 
 	: m_radius { r }	{
 		add_point(p);
 	}
 
-	Circle(int x, int y, int r) : Circle { {x, y}, r} { };
+	Circle(float x, float y, int r) : Circle { {x, y}, r} { };
 	Circle() 
 		: Circle({0, 0}, 1) { };
 
@@ -20,13 +20,13 @@ public:
 	void compute_fill(Frame & f);
 
 	// getters
-	V2d<int> pos() const { return get_point(0); };
-	int radius() 	const { return m_radius; };
+	V2d<float> pos() const { return get_point(0); };
+	int radius() 	  const { return m_radius; };
 	
 	// setters
-	void pos(V2d<int> p) 	{ set_point(0, p); };
-	void pos(int x, int y) 	{ set_point(0, V2d<int> {x, y}); };
-	void radius(int r) 		{ m_radius = r; };
+	void pos(V2d<float> p) 		{ set_point(0, p); };
+	void pos(float x, float y) { set_point(0, V2d<float> {x, y}); };
+	void radius(int r) 			{ m_radius = r; };
 
 private:
 	int m_radius;
